+++ tests/compas/geometry/test_polyline.py
@@ -122,7 +122,7 @@ def test_polyline_tangent_at_point(coords, input, expected):
     ([[0, 0, 0], [1, 0, 0], [2, 0, 0], [2, 2, 0]], (2, 0), [[-2, 0, 0], [1, 0, 0], [2, 0, 0], [2, 2, 0]]),
 ])
 def test_polyline_extend(coords, input, expected):
-    assert expected == Polyline(coords).extend(input)
+    assert expected == Polyline(coords).extended(input)
 
 
 @pytest.mark.parametrize('coords,input,expected', [
@@ -132,5 +132,5 @@ def test_polyline_extend(coords, input, expected):
     ([[0, 0, 0], [1, 0, 0], [2, 0, 0], [2, 2, 0]], (1, 2), [[1, 0, 0], [2, 0, 0]]),
 ])
 def test_polyline_extend(coords, input, expected):
-    assert expected == Polyline(coords).shorten(input)
+    assert expected == Polyline(coords).shortened(input)